fix some x64 warnings
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@1495 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/subtitles/Tfont.cpp
@@ -1673,7 +1673,7 @@ template<class tchar> void Tfont::prepareC(const TsubtitleTextBase<tchar> *sub,c
                p++;
               }
             }
-           int strlenp=strlen(p);
+           int strlenp=(int)strlen(p);
            if (cx+strlenp-1<=wordWrap.getRightOfTheLine(cy))
             {
              if (*p)
